removed game app
