--- sdk/freertos_app_cpu0/src/usr/game/app_game.h
@@ -1,6 +1,0 @@
-#ifndef APP_GAME_H
-#define APP_GAME_H
-
-void app_game_init(void);
-
-#endif // APP_GAME_H
