--- sdk/freertos_app_cpu0/src/usr/game/cmd/cmd_game.h
@@ -1,12 +1,0 @@
-#ifndef CMD_GAME_H
-#define CMD_GAME_H
-
-void cmd_game_register(void);
-
-int cmd_game1(int argc, char **argv);
-int cmd_game2(int argc, char **argv);
-int cmd_game3(int argc, char **argv);
-int cmd_game4(int argc, char **argv);
-int cmd_gameRestart(int argc, char **argv);
-
-#endif // CMD_GAME_H
